Disaggregated expression results did not exactly match results from original expression. (Lahey Fortran 95 compiler) Likely cause is that original expression retained intermediate values in registers (at higher precision) while disaggregated expression used intermediate variables (at standard predision). Increasing precision of intermediate variables made the results between disaggregated and original expression match exactly at standard precision.
git-svn-id: https://infosys.ars.usda.gov/svn/code/weps1/branches/weps.src.subregion@12769 77af8a48-1994-4e4a-b6f4-b1c89642997b

+++ src/lib_erosion/process_mod.f95
@@ -497,8 +497,8 @@ module process_mod
 
       real :: sargc    ! soil angle of ridge shelter weibull 'c' (deg.)
       real :: sarrc    ! soil angle of random roughness weibull 'c'(deg.)
-      real :: sfrg12   ! sheltered soil fraction for ridge
-      real :: sfrr12   ! sheltered soil fraction for random roughness
+      double precision :: sfrg12   ! sheltered soil fraction for ridge
+      double precision :: sfrr12   ! sheltered soil fraction for random roughness
 
       sarrc = 2.3 * sqrt (slrr)
       sfrr12 = exp(-(12.0/sarrc)**0.77)
